002.6.1. Реализовать авторизация и аутентификацию[#251562]

+++ servlet/src/main/java/ru/job4j/servlets/controllers/CreateUserController.java
@@ -1,15 +1,18 @@
-package ru.job4j.servlets;
+package ru.job4j.servlets.controllers;
 
 import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import ru.job4j.servlets.UserStorage;
+import ru.job4j.servlets.ValidateService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -19,10 +22,12 @@ import java.util.List;
 
 public class CreateUserController extends HttpServlet {
     private final ValidateService service = ValidateService.getInstance();
+    private final UserStorage userStorage = UserStorage.getInstance();
     private static final Logger log = LogManager.getLogger(CreateUserController.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.setAttribute("roles", userStorage.getUserRoles());
         req.getRequestDispatcher("/WEB-INF/views/create.jsp").forward(req, resp);
     }
 
@@ -34,11 +39,12 @@ public class CreateUserController extends HttpServlet {
             resp.sendError(HttpServletResponse.SC_BAD_REQUEST, "Not Multipart Content");
             return;
         }
-        req.setCharacterEncoding("utf-8");
         String name = null;
         String login = null;
+        String password = null;
         String email = null;
         String imageFile = null;
+        String role = null;
         DiskFileItemFactory factory = new DiskFileItemFactory();
         File tempDir = (File) getServletContext().getAttribute("javax.servlet.context.tempdir");
         factory.setRepository(tempDir);
@@ -49,29 +55,44 @@ public class CreateUserController extends HttpServlet {
             List<FileItem> items = upload.parseRequest(req);
             for (FileItem item : items) {
                 if (item.isFormField()) {
-                    name = "name".equals(item.getFieldName()) ? item.getString() : name;
-                    login = "login".equals(item.getFieldName()) ? item.getString() : login;
-                    email = "email".equals(item.getFieldName()) ? item.getString() : email;
+                    name = "name".equals(item.getFieldName()) ? item.getString("utf-8") : name;
+                    login = "login".equals(item.getFieldName()) ? item.getString("utf-8") : login;
+                    password = "password".equals(item.getFieldName()) ? item.getString("utf-8") : password;
+                    email = "email".equals(item.getFieldName()) ? item.getString("utf-8") : email;
+                    role = "role".equals(item.getFieldName()) ? item.getString("utf-8") : role;
                 } else {
                     imageFile = item.getName();
-                    File uploadedFile = new File(path + imageFile);
-                    uploadedFile.mkdirs();
-                    if (uploadedFile.exists()) {
-                        uploadedFile.delete();
+                    if (!imageFile.isEmpty()) {
+                        File uploadedFile = new File(path + imageFile);
+                        uploadedFile.mkdirs();
+                        if (uploadedFile.exists()) {
+                            uploadedFile.delete();
+                        }
+                        item.write(uploadedFile);
                     }
-                    item.write(uploadedFile);
                 }
             }
-            long id = service.addUser(name, login, email, imageFile);
-            if (id != -1) {
-                Files.move(Path.of(path + imageFile), Path.of(path + id + "_" + imageFile), StandardCopyOption.REPLACE_EXISTING);
+            long id = service.addUser(name, login, password, email, imageFile, role);
+            if (id == -2) {
+                String message = String.format("User with login %s already exists", login);
+                log.trace(message);
+                resp.sendError(HttpServletResponse.SC_CONFLICT, message);
+                return;
+            }
+            if (id > 0) {
+                renameImageFile(String.valueOf(id), path, imageFile);
             }
         } catch (Exception e) {
             log.error(e.getMessage(), e);
-            resp.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+            resp.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, e.getMessage());
             return;
         }
         resp.sendRedirect("/");
+    }
+    private void renameImageFile(String id, String path, String imageFile) throws IOException {
+        if (imageFile != null && !imageFile.isEmpty()) {
+            Files.move(Path.of(path + imageFile), Path.of(path + id + "_" + imageFile), StandardCopyOption.REPLACE_EXISTING);
+        }
     }
 }
